+++ src/ReadRSEMBias.cpp
@@ -55,7 +55,7 @@ vector<double> ReadRSEMrspd(string filename)
 		if (is_rspd_line) {
 			vector<string> strs;
 			boost::split(strs, line, boost::is_any_of("\t "));
-			assert( strs.size() == stoi(prevline) );
+			assert((int32_t)strs.size() == stoi(prevline) );
 
 			for (uint32_t i = 0; i < strs.size(); i++)
 				rspd.push_back( stod(strs[i]) );
@@ -103,7 +103,7 @@ vector<double> BiasCorrectTrans(string seq, const vector<double>& rspd)
 			idx += 1;
 			start = (int32_t)round(1.0 * idx * seq.size() / rspd.size());
 			end = (int32_t)round(1.0 * (idx + 1) * seq.size() / rspd.size());
-			assert( end <= seq.size() );
+			assert(end <= (int32_t)seq.size() );
 			double tmpsum = std::accumulate(correction.begin(), correction.begin() + i, 0.0);
 			double tmptruth = std::accumulate(rspd.begin(), rspd.begin() + idx, 0.0);
 			if (fabs(tmpsum - tmptruth) > 1e-8)
